+++ packages/liveblocks-core/src/index.ts
@@ -267,6 +267,7 @@ export { WebsocketCloseCodes } from "./types/IWebSocket";
 export type { NodeMap, ParentToChildNodeMap } from "./types/NodeMap";
 export type { OptionalPromise } from "./types/OptionalPromise";
 export type { OthersEvent } from "./types/Others";
+export { TextEditorType } from "./types/Others";
 export type { Patchable } from "./types/Patchable";
 export type {
   PlainLson,